feat: proposal for server agnostic handlers. Refactor
