+++ packages/openid4vc/tests/openid4vc-batch-issuance.e2e.test.ts
@@ -83,7 +83,10 @@ describe('OpenId4Vc Presentation During Issuance', () => {
     await issuer.agent.x509.addTrustedCertificate(issuer.certificate.toString('base64'))
 
     // We let AFJ create the router, so we have a fresh one each time
-    expressApp.use('/oid4vci', issuer.agent.modules.openId4VcIssuer.config.router)
+    expressApp.use(
+      '/oid4vci',
+      (issuer.agent.dependencyManager.registeredModules['openId4VcIssuer'] as OpenId4VcIssuerModule).contextRouter
+    )
     clearNock = setupNockToExpress(baseUrl, expressApp)
   })
