+++ packages/openid4vc/tests/openid4vc-presentation-during-issuance.e2e.test.ts
@@ -161,7 +161,10 @@ describe('OpenId4Vc Presentation During Issuance', () => {
     issuer.agent.x509.addTrustedCertificate(issuer.certificate.toString('base64'))
 
     // We let AFJ create the router, so we have a fresh one each time
-    expressApp.use('/oid4vci', issuer.agent.modules.openId4VcIssuer.config.router)
+    expressApp.use(
+      '/oid4vci',
+      (issuer.agent.dependencyManager.registeredModules['openId4VcIssuer'] as OpenId4VcIssuerModule).contextRouter
+    )
     expressApp.use('/oid4vp', issuer.agent.modules.openId4VcVerifier.config.router)
 
     clearNock = setupNockToExpress(baseUrl, expressApp)
